fix precision loss while using depth strider

+++ common/src/main/java/ac/grim/grimac/predictionengine/movementtick/MovementTicker.java
@@ -430,7 +430,7 @@ public class MovementTicker {
             if (player.depthStriderLevel > 0.0F) {
                 final float divisor = player.getClientVersion().isNewerThanOrEquals(ClientVersion.V_1_21) ? 1.0F : 3.0F;
                 swimFriction += (0.54600006F - swimFriction) * player.depthStriderLevel / divisor;
-                swimSpeed += (player.speed - swimSpeed) * player.depthStriderLevel / divisor;
+                swimSpeed += (((float) player.speed) - swimSpeed) * player.depthStriderLevel / divisor;
             }
 
             if (player.compensatedEntities.getPotionLevelForPlayer(PotionTypes.DOLPHINS_GRACE).isPresent()) {
